Shift Nix offsets minus one, from host-target to build-host
Rust uses host-target terminology.  From the perspective of serving toolchains,
this makes sense.  You are on a build platform and distributing toolchains to
run on hosts and generate outputs for targets.

However, from the perspective of the person downloading a toolchain, they are
seeking a compiler for their buildPlatform to generate outputs for their
hostPlatform, where their Rust program will run.

This change shifts most targets to host and hosts to build.

When providing build inputs for a native or cross build, depsBuildHost are
available during build time to make outputs for the host.  Therefore, we
propagate libiconv for linking into the outputs on the host.  depsBuildHost is
also known as nativeBuildInputs.  propagatedNativeBuildInputs could also be used
here.

https://nixos.org/manual/nixpkgs/stable/#variables-specifying-dependencies