mk-component-sets: add `meta.mainProgram` for components (#204)
Previously the individual components have not set a mainProgram. This is
the somewhat de-facto standard in nixpkgs to signal which executable is
associated to a given derivation.

This patch uses the renames table for the given components to get the
proper executable name.

To be conservative, only these executables are marked for now. They are
also frequently run on their own without cargo.

Signed-off-by: Marcel Müller <neikos@neikos.email>